requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SocketIO==5.3.5
python-socketio==5.10.0
requests==2.31.0
pandas==2.1.4
numpy==1.26.2
python-dotenv==1.0.0
SQLAlchemy==2.0.23
redis==5.0.1
websocket-client==1.7.0
eventlet==0.35.1
scipy==1.11.4